Refactor basedecorator to be on its own file
