+++ decorator.rb
@@ -1,0 +1,14 @@
+require_relative './interface'
+
+class BaseDecorator < Nameable
+    attr_accessor :nameable
+  
+    def initialize(nameable)
+      super()
+      @nameable = nameable
+    end
+  
+    def correct_name
+      @nameable.correct_name
+    end
+  end
