+++ interface.rb
@@ -4,18 +4,6 @@ class Nameable
   end
 end
 
-class BaseDecorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
 
 class CapitalizeDecorator < BaseDecorator
   def correct_name
